Added change password handler
