+++ ws/src/main/java/handlers/ClientProfileHandler.kt
@@ -9,6 +9,7 @@ import foi.air.core.models.PhysicalMeasurements
 import foi.air.core.models.TargetMeasurement
 import foi.air.core.models.UserData
 import foi.air.core.models.UserDataResponse
+import org.json.JSONObject
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -32,8 +33,10 @@ class DefaultClientProfileHandler : ClientProfileHandler {
                     val user: UserData? = responseData?.data
                     callback.invoke(true, user, null)
                 } else {
-                    val error = response.errorBody()?.string() ?: "Unknown error"
-                    callback.invoke(false, null, error)
+                    val errorResponse = response.errorBody()?.string() ?: "Unknown error"
+                    val errorJson = JSONObject(errorResponse)
+                    val errorMessage = errorJson.optString("message", "Unknown error")
+                    callback.invoke(false, null, errorMessage)
                 }
             }
 
@@ -55,8 +58,10 @@ class DefaultClientProfileHandler : ClientProfileHandler {
                     val physicalMeasurements: List<PhysicalMeasurements>? = responseMeasurementData?.data?.physical_measurements
                     callback.invoke(true, responseMeasurementData, null)
                 } else {
-                    val error = response.errorBody()?.string() ?: "Unknown error"
-                    callback.invoke(false, null, error)
+                    val errorResponse = response.errorBody()?.string() ?: "Unknown error"
+                    val errorJson = JSONObject(errorResponse)
+                    val errorMessage = errorJson.optString("message", "Unknown error")
+                    callback.invoke(false, null, errorMessage)
                 }
             }
